Moved create_sumo_client() factory function from _helpers.py into separate file sumo_client_factory.py
