+++ backend_py/primary/primary/services/sumo_access/observation_access.py
@@ -7,7 +7,8 @@ from fmu.sumo.explorer.objects.dictionary import Dictionary
 
 from primary.services.service_exceptions import InvalidDataError, MultipleDataMatchesError, Service
 
-from ._helpers import create_sumo_client, create_sumo_case_async
+from ._helpers import create_sumo_case_async
+from .sumo_client_factory import create_sumo_client
 from .observation_types import (
     Observations,
     SummaryVectorDateObservation,
