+++ backend_py/primary/primary/services/sumo_access/parameter_access.py
@@ -10,7 +10,8 @@ from webviz_pkg.core_utils.perf_timer import PerfTimer
 
 from primary.services.service_exceptions import InvalidDataError, MultipleDataMatchesError, NoDataError, Service
 
-from ._helpers import create_sumo_client, create_sumo_case_async
+from ._helpers import create_sumo_case_async
+from .sumo_client_factory import create_sumo_client
 from .parameter_types import (
     EnsembleParameter,
     EnsembleParameters,
